+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Webhooks (beta) Changelog
 
+## 2nd April 2024
+
+- Added proof overdue webhook.
+
 ## 12th February 2024
 
 - Added proof deleted webhook.